lógica de impressão de etiqueta e vale
lógica de impressão de etiqueta e vale
